superstock_analyzer_v2/requirements.txt
yfinance>=0.2.3
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=0.27.0
beautifulsoup4>=4.9.3
matplotlib>=3.5.0
seaborn>=0.11.0
schedule>=1.1.0
scipy>=1.7.0
ta-lib>=0.4.24
aiohttp>=3.8.0
aiosignal>=1.3.1
